+++ src/diagnostics.rs
@@ -22,6 +22,8 @@ pub struct DiagnosticsConfig {
   pub unused_definition: Severity,
   pub repeated_bind: Severity,
   pub recursion_cycle: Severity,
+  pub missing_main: Severity,
+  pub import_shadow: Severity,
 }
 
 #[derive(Debug, Clone)]
@@ -57,6 +59,8 @@ pub enum WarningType {
   UnusedDefinition,
   RepeatedBind,
   RecursionCycle,
+  MissingMain,
+  ImportShadow,
 }
 
 impl Diagnostics {
@@ -235,6 +239,9 @@ impl DiagnosticsConfig {
       unused_definition: severity,
       repeated_bind: severity,
       recursion_cycle: severity,
+      import_shadow: severity,
+      // Should only be changed manually, as a missing main is always a error to hvm
+      missing_main: Severity::Error,
       verbose,
     }
   }
@@ -247,6 +254,8 @@ impl DiagnosticsConfig {
       WarningType::IrrefutableMatch => self.irrefutable_match,
       WarningType::RedundantMatch => self.redundant_match,
       WarningType::UnreachableMatch => self.unreachable_match,
+      WarningType::MissingMain => self.missing_main,
+      WarningType::ImportShadow => self.import_shadow,
     }
   }
 }