+++ src/fun/parser.rs
@@ -2,14 +2,60 @@ use std::ops::Range;
 
 use crate::{
   fun::{
-    display::DisplayFn, Adt, Book, CtrField, Definition, FanKind, HvmDefinition, MatchRule, Name, Num, Op,
-    Pattern, Rule, Tag, Term, STRINGS,
+    display::DisplayFn, Adt, Adts, Constructors, CtrField, FanKind, HvmDefinition, HvmDefinitions, MatchRule,
+    Name, Num, Op, Pattern, Rule, Source, Tag, Term, STRINGS,
   },
   imp::{parser::PyParser, Enum, RepeatedNames, Variant},
+  imports::{Import, ImportCtx, ImportType},
   maybe_grow,
 };
 use highlight_error::highlight_error;
+use indexmap::IndexMap;
+use itertools::Itertools;
 use TSPL::Parser;
+
+type FunDefinition = super::Definition;
+type ImpDefinition = crate::imp::Definition;
+
+/// Intermediate representation of a program.
+#[derive(Debug, Clone, Default)]
+pub struct ParseBook {
+  /// The `functional` function definitions.
+  pub fun_defs: IndexMap<Name, FunDefinition>,
+
+  /// The `imperative` function definitions.
+  pub imp_defs: IndexMap<Name, ImpDefinition>,
+
+  /// HVM native function definitions.
+  pub hvm_defs: HvmDefinitions,
+
+  /// The algebraic datatypes defined by the program
+  pub adts: Adts,
+
+  /// To which type does each constructor belong to.
+  pub ctrs: Constructors,
+
+  /// Imported packages to be loaded in the program
+  pub import_ctx: ImportCtx,
+
+  /// Source of the book
+  pub source: Name,
+}
+
+impl ParseBook {
+  pub fn contains_def(&self, name: &Name) -> bool {
+    self.fun_defs.contains_key(name) || self.imp_defs.contains_key(name) || self.hvm_defs.contains_key(name)
+  }
+
+  pub fn contains_builtin_def(&self, name: &Name) -> Option<bool> {
+    self
+      .fun_defs
+      .get(name)
+      .map(|d| d.is_builtin())
+      .or_else(|| self.imp_defs.get(name).map(|d| d.source.is_builtin()))
+      .or_else(|| self.hvm_defs.get(name).map(|d| d.source.is_builtin()))
+  }
+}
 
 // Bend grammar description:
 // <Book>       ::= (<Data> | <Rule>)*
@@ -65,7 +111,7 @@ impl<'a> TermParser<'a> {
 
   /* AST parsing functions */
 
-  pub fn parse_book(&mut self, default_book: Book, builtin: bool) -> ParseResult<Book> {
+  pub fn parse_book(&mut self, default_book: ParseBook, builtin: bool) -> ParseResult<ParseBook> {
     let mut book = default_book;
     let mut indent = self.advance_newlines()?;
     let mut last_rule = None;
@@ -116,8 +162,10 @@ impl<'a> TermParser<'a> {
         // Fun type definition
         } else {
           self.index = rewind_index;
-          let (nam, adt) = self.parse_datatype(builtin)?;
+          let (nam, ctrs) = self.parse_datatype()?;
           let end_idx = *self.index();
+          let source = if builtin { Source::Builtin } else { Source::Local(ini_idx..end_idx) };
+          let adt = Adt { ctrs, source };
           self.add_fun_type(&mut book, nam, adt, ini_idx..end_idx)?;
           indent = self.advance_newlines()?;
           last_rule = None;
@@ -135,10 +183,37 @@ impl<'a> TermParser<'a> {
         continue;
       }
 
+      // Import declaration
+      if self.try_parse_keyword("from") {
+        self.skip_trivia();
+        let import = self.parse_from_import()?;
+        book.import_ctx.add_import(import);
+        indent = self.advance_newlines()?;
+        last_rule = None;
+        continue;
+      }
+
+      if self.try_parse_keyword("import") {
+        self.skip_trivia();
+        let imports = self.parse_import()?;
+        for imp in imports {
+          book.import_ctx.add_import(imp);
+        }
+        indent = self.advance_newlines()?;
+        last_rule = None;
+        continue;
+      }
+
       // Fun function definition
       let ini_idx = *self.index();
       let (name, rule) = self.parse_rule()?;
       let end_idx = *self.index();
+
+      if let Some(def) = book.imp_defs.get(&name) {
+        let msg = Self::redefinition_of_function_msg(def.source.is_builtin(), &name);
+        return self.with_ctx(Err(msg), ini_idx..end_idx);
+      }
+
       self.add_fun_def(&name, rule, builtin, &last_rule, &mut book, ini_idx..end_idx)?;
       indent = self.advance_newlines()?;
       last_rule = Some(name);
@@ -147,7 +222,7 @@ impl<'a> TermParser<'a> {
     Ok(book)
   }
 
-  fn parse_datatype(&mut self, builtin: bool) -> ParseResult<(Name, Adt)> {
+  fn parse_datatype(&mut self) -> ParseResult<(Name, IndexMap<Name, Vec<CtrField>>)> {
     // type name = ctr (| ctr)*
     self.skip_trivia();
     let name = self.labelled(|p| p.parse_top_level_name(), "datatype name")?;
@@ -157,8 +232,7 @@ impl<'a> TermParser<'a> {
       ctrs.push(self.parse_datatype_ctr(&name)?);
     }
     let ctrs = ctrs.into_iter().collect();
-    let adt = Adt { ctrs, builtin };
-    Ok((name, adt))
+    Ok((name, ctrs))
   }
 
   fn parse_datatype_ctr(&mut self, typ_name: &Name) -> ParseResult<(Name, Vec<CtrField>)> {
@@ -200,8 +274,56 @@ impl<'a> TermParser<'a> {
     let mut p = hvm::ast::CoreParser::new(&self.input[*self.index()..]);
     let body = p.parse_net()?;
     *self.index() = ini_idx + *p.index();
-    let def = HvmDefinition { name: name.clone(), body, builtin };
+    let end_idx = *self.index();
+    let source = if builtin { Source::Builtin } else { Source::Local(ini_idx..end_idx) };
+    let def = HvmDefinition { name: name.clone(), body, source };
     Ok(def)
+  }
+
+  fn parse_from_import(&mut self) -> Result<Import, String> {
+    // from path import package
+    // from path import (a, b)
+    // from path import *
+    let path = self.parse_restricted_name("Path")?;
+    self.consume("import")?;
+
+    let relative = path.starts_with("./") | path.starts_with("../");
+
+    if self.try_consume("*") {
+      return Ok(Import::new(path, ImportType::Glob, relative));
+    }
+
+    if self.try_consume("(") {
+      let sub = self.list_like(|p| p.parse_name_maybe_alias("Name"), "", ")", ",", false, 1)?;
+      return Ok(Import::new(path, ImportType::List(sub), relative));
+    }
+
+    let (import, alias) = self.parse_name_maybe_alias("Import")?;
+    Ok(Import::new(path, ImportType::Single(import, alias), relative))
+  }
+
+  fn parse_import(&mut self) -> Result<Vec<Import>, String> {
+    // import path
+    // import (path/a, path/b)
+
+    let new_import = |import: Name, alias: Option<Name>, relative: bool| -> Import {
+      let (path, import) = match import.rsplit_once('/') {
+        Some((start, end)) => (Name::new(start), Name::new(end)),
+        None => (Name::default(), import),
+      };
+
+      Import::new(path, ImportType::Single(import, alias), relative)
+    };
+
+    if self.try_consume("(") {
+      let list = self.list_like(|p| p.parse_import_name("Name"), "", ")", ",", false, 1)?;
+      let imports = list.into_iter().map(|(a, b, c)| new_import(a, b, c)).collect_vec();
+      return Ok(imports);
+    }
+
+    let (import, alias, relative) = self.parse_import_name("Import")?;
+    let import = new_import(import, alias, relative);
+    Ok(vec![import])
   }
 
   fn parse_rule(&mut self) -> ParseResult<(Name, Rule)> {
@@ -783,28 +905,32 @@ impl<'a> TermParser<'a> {
     rule: Rule,
     builtin: bool,
     last_rule: &Option<Name>,
-    book: &mut Book,
+    book: &mut ParseBook,
     span: Range<usize>,
   ) -> ParseResult<()> {
-    match (book.defs.get_mut(name), last_rule) {
+    match (book.fun_defs.get_mut(name), last_rule) {
       // Continuing with a new rule to the current definition
       (Some(def), Some(last_rule)) if last_rule == name => {
         def.rules.push(rule);
+        if let Source::Local(s) = &mut def.source {
+          s.end = span.end;
+        }
       }
       // Trying to add a new rule to a previous definition, coming from a different rule.
       (Some(def), Some(_)) => {
-        let msg = Self::redefinition_of_function_msg(def.builtin, name);
+        let msg = Self::redefinition_of_function_msg(def.is_builtin(), name);
         return self.with_ctx(Err(msg), span);
       }
       // Trying to add a new rule to a previous definition, coming from another kind of top-level.
       (Some(def), None) => {
-        let msg = Self::redefinition_of_function_msg(def.builtin, name);
+        let msg = Self::redefinition_of_function_msg(def.is_builtin(), name);
         return self.with_ctx(Err(msg), span);
       }
       // Adding the first rule of a new definition
       (None, _) => {
-        self.check_top_level_redefinition(name, book, span)?;
-        book.defs.insert(name.clone(), Definition { name: name.clone(), rules: vec![rule], builtin });
+        self.check_top_level_redefinition(name, book, span.clone())?;
+        let source = if builtin { Source::Builtin } else { Source::Local(span) };
+        book.fun_defs.insert(name.clone(), FunDefinition::new(name.clone(), vec![rule], source));
       }
     }
     Ok(())
@@ -813,19 +939,18 @@ impl<'a> TermParser<'a> {
   fn add_imp_def(
     &mut self,
     mut def: crate::imp::Definition,
-    book: &mut Book,
+    book: &mut ParseBook,
     span: Range<usize>,
     builtin: bool,
   ) -> ParseResult<()> {
-    self.check_top_level_redefinition(&def.name, book, span)?;
-    def.order_kwargs(book)?;
-    def.gen_map_get();
-    let def = def.to_fun(builtin)?;
-    book.defs.insert(def.name.clone(), def);
+    self.check_top_level_redefinition(&def.name, book, span.clone())?;
+    let source = if builtin { Source::Builtin } else { Source::Local(span) };
+    def.source = source;
+    book.imp_defs.insert(def.name.clone(), def);
     Ok(())
   }
 
-  fn add_hvm(&mut self, def: HvmDefinition, book: &mut Book, span: Range<usize>) -> ParseResult<()> {
+  fn add_hvm(&mut self, def: HvmDefinition, book: &mut ParseBook, span: Range<usize>) -> ParseResult<()> {
     self.check_top_level_redefinition(&def.name, book, span)?;
     book.hvm_defs.insert(def.name.clone(), def);
     Ok(())
@@ -834,12 +959,13 @@ impl<'a> TermParser<'a> {
   fn add_imp_type(
     &mut self,
     enum_: Enum,
-    book: &mut Book,
+    book: &mut ParseBook,
     span: Range<usize>,
     builtin: bool,
   ) -> ParseResult<()> {
     self.check_type_redefinition(&enum_.name, book, span.clone())?;
-    let mut adt = Adt { ctrs: Default::default(), builtin };
+    let source = if builtin { Source::Builtin } else { Source::Local(span.clone()) };
+    let mut adt = Adt { ctrs: Default::default(), source };
     for variant in enum_.variants {
       self.check_top_level_redefinition(&enum_.name, book, span.clone())?;
       book.ctrs.insert(variant.name.clone(), enum_.name.clone());
@@ -849,12 +975,21 @@ impl<'a> TermParser<'a> {
     Ok(())
   }
 
-  fn add_fun_type(&mut self, book: &mut Book, nam: Name, adt: Adt, span: Range<usize>) -> ParseResult<()> {
+  fn add_fun_type(
+    &mut self,
+    book: &mut ParseBook,
+    nam: Name,
+    adt: Adt,
+    span: Range<usize>,
+  ) -> ParseResult<()> {
     if book.adts.contains_key(&nam) {
       let msg = TermParser::redefinition_of_type_msg(&nam);
       return self.with_ctx(Err(msg), span);
     } else {
       for ctr in adt.ctrs.keys() {
+        if let Some(builtin) = book.contains_builtin_def(ctr) {
+          return Err(TermParser::redefinition_of_function_msg(builtin, ctr));
+        }
         match book.ctrs.entry(ctr.clone()) {
           indexmap::map::Entry::Vacant(e) => _ = e.insert(nam.clone()),
           indexmap::map::Entry::Occupied(e) => {
@@ -871,13 +1006,14 @@ impl<'a> TermParser<'a> {
   fn add_object(
     &mut self,
     obj: Variant,
-    book: &mut Book,
+    book: &mut ParseBook,
     span: Range<usize>,
     builtin: bool,
   ) -> ParseResult<()> {
     self.check_type_redefinition(&obj.name, book, span.clone())?;
-    self.check_top_level_redefinition(&obj.name, book, span)?;
-    let mut adt = Adt { ctrs: Default::default(), builtin };
+    self.check_top_level_redefinition(&obj.name, book, span.clone())?;
+    let source = if builtin { Source::Builtin } else { Source::Local(span) };
+    let mut adt = Adt { ctrs: Default::default(), source };
     book.ctrs.insert(obj.name.clone(), obj.name.clone());
     adt.ctrs.insert(obj.name.clone(), obj.fields);
     book.adts.insert(obj.name, adt);
@@ -887,11 +1023,11 @@ impl<'a> TermParser<'a> {
   fn check_top_level_redefinition(
     &mut self,
     name: &Name,
-    book: &mut Book,
+    book: &mut ParseBook,
     span: Range<usize>,
   ) -> ParseResult<()> {
-    if let Some(def) = book.defs.get(name) {
-      let msg = Self::redefinition_of_function_msg(def.builtin, name);
+    if let Some(builtin) = book.contains_builtin_def(name) {
+      let msg = Self::redefinition_of_function_msg(builtin, name);
       return self.with_ctx(Err(msg), span);
     }
     if book.ctrs.contains_key(name) {
@@ -905,7 +1041,12 @@ impl<'a> TermParser<'a> {
     Ok(())
   }
 
-  fn check_type_redefinition(&mut self, name: &Name, book: &mut Book, span: Range<usize>) -> ParseResult<()> {
+  fn check_type_redefinition(
+    &mut self,
+    name: &Name,
+    book: &mut ParseBook,
+    span: Range<usize>,
+  ) -> ParseResult<()> {
     if book.adts.contains_key(name) {
       let msg = Self::redefinition_of_type_msg(name);
       return self.with_ctx(Err(msg), span);
@@ -999,8 +1140,6 @@ impl Indent {
   }
 }
 
-impl Book {}
-
 impl<'a> ParserCommons<'a> for TermParser<'a> {}
 
 pub trait ParserCommons<'a>: Parser<'a> {
@@ -1036,6 +1175,24 @@ pub trait ParserCommons<'a>: Parser<'a> {
 
   fn parse_bend_name(&mut self) -> ParseResult<Name> {
     self.parse_restricted_name("Variable")
+  }
+
+  fn parse_name_maybe_alias(&mut self, label: &str) -> ParseResult<(Name, Option<Name>)> {
+    let name = self.parse_restricted_name(label)?;
+
+    if self.try_consume("as") {
+      self.skip_trivia();
+      let alias = self.parse_restricted_name("Alias")?;
+      Ok((name, Some(alias)))
+    } else {
+      Ok((name, None))
+    }
+  }
+
+  fn parse_import_name(&mut self, label: &str) -> Result<(Name, Option<Name>, bool), String> {
+    let (import, alias) = self.parse_name_maybe_alias(label)?;
+    let relative = import.starts_with("./") | import.starts_with("../");
+    Ok((import, alias, relative))
   }
 
   /// Consumes exactly the text without skipping.