+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project does not currently adhere to a particular versioning scheme.
 
+## [Unreleased]
+
+### Fixed
+
+- Fix variable binding in pattern matching when the irrefutable pattern optimization occurs. ([#618][gh-618])
+- Don't warn on unused generated definitions. ([#514][gh-514])
+
+### Added
+
+- Add import system ([#544][gh-544])
+
 ## [0.2.36] - 2024-07-04
 
 ### Changed
@@ -36,6 +47,7 @@ and this project does not currently adhere to a particular versioning scheme.
 - Add local definitions to imperative syntax. ([#562][gh-562])
 - Add local definitions to functional syntax.
 - Add repeated field name error message.
+- Add `Math` builtin functions. ([#570][gh-570])
 - Add primitive file IO function `IO/FS/flush`. ([#598][gh-598])
 
 ## [0.2.35] - 2024-06-06
@@ -336,6 +348,7 @@ and this project does not currently adhere to a particular versioning scheme.
 [0.2.34]: https://github.com/HigherOrderCO/Bend/releases/tag/0.2.34
 [0.2.35]: https://github.com/HigherOrderCO/Bend/releases/tag/0.2.35
 [0.2.36]: https://github.com/HigherOrderCO/Bend/releases/tag/0.2.36
+[Unreleased]: https://github.com/HigherOrderCO/Bend/compare/0.2.36...HEAD
 [gh-424]: https://github.com/HigherOrderCO/Bend/issues/424
 [gh-427]: https://github.com/HigherOrderCO/Bend/issues/427
 [gh-443]: https://github.com/HigherOrderCO/Bend/issues/443
@@ -361,13 +374,18 @@ and this project does not currently adhere to a particular versioning scheme.
 [gh-526]: https://github.com/HigherOrderCO/Bend/issues/526
 [gh-528]: https://github.com/HigherOrderCO/Bend/issues/528
 [gh-562]: https://github.com/HigherOrderCO/Bend/issues/562
-[gh-581]: https://github.com/HigherOrderCO/Bend/issues/581
+[gh-570]: https://github.com/HigherOrderCO/Bend/issues/570
 [gh-573]: https://github.com/HigherOrderCO/Bend/issues/573
 [gh-578]: https://github.com/HigherOrderCO/Bend/issues/578
 [gh-579]: https://github.com/HigherOrderCO/Bend/issues/579
 [gh-580]: https://github.com/HigherOrderCO/Bend/issues/580
+[gh-581]: https://github.com/HigherOrderCO/Bend/issues/581
 [gh-582]: https://github.com/HigherOrderCO/Bend/issues/582
 [gh-583]: https://github.com/HigherOrderCO/Bend/issues/583
 [gh-586]: https://github.com/HigherOrderCO/Bend/issues/586
 [gh-596]: https://github.com/HigherOrderCO/Bend/issues/596
+[gh-618]: https://github.com/HigherOrderCO/Bend/issues/618
+[gh-514]: https://github.com/HigherOrderCO/Bend/issues/514
+[gh-544]: https://github.com/HigherOrderCO/Bend/pull/544
+[Unreleased]: https://github.com/HigherOrderCO/Bend/compare/0.2.36...HEAD
 [gh-598]: https://github.com/HigherOrderCO/Bend/issues/598